Display Photos in CollectionView using NSFetchResultsController
The Photo Album View Controller checks if the image exists in the
Documents Directory before downloading it using the imageURL. A
placeholder image is displayed until the new image is downloaded. The
downloaded image is then saved in the Documents Directory. The photo id
is used to name the file uniquely.